requirements.txt
asgiref==3.5.2
backports.zoneinfo==0.2.1
Django==4.1.2
djangorestframework==3.14.0
importlib-metadata==5.0.0
Markdown==3.4.1
pytz==2022.6
sqlparse==0.4.3
zipp==3.9.0
